backend/api/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from api.views import UserViewSet

app_name = 'api'

router = DefaultRouter()
router.register('users', UserViewSet, 'users')

urlpatterns = (
    path('auth/', include('djoser.urls.authtoken')),
    path('', include(router.urls)),
)
