backend/requirements.txt
django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0